Multiprocessing tests
Since the "DownNote()" routine has delays in it; it would have to wait for the next event for a new note (meaning that you couldn't use special high level effects and do something else on another channel). but you can if you place that sub-routine on it's own cog. I will still have to experiment with the optimal stack space. The stack will likely be channel based and not subroutine based, but we will see.